fix: testerfeedback runde 1 — sju retta punkter
- «Legg til oevelse virker ikke»: oevelsen BLE lagt til, men bakerst —
  single-oevelse-pageren ble staaende. Hopper naa til den nye oevelsen.
- Hopp-til-sikting: blokk-bytte scroller naa oevelseskortets topp til
  toppen av skjermen (foer beholdt den gammel scrollposisjon, saa kortet
  kunne starte langt nede).
- Rest-innstillingene: innholdet er naa scrollbart (ble klippet av
  maxHeight), KeyboardAvoidingView saa tastaturet ikke gjemmer
  sekund-feltene, forklaringstekst oeverst om hva timeren gjoer, og
  preset-chips vises som m:ss (2:00) i stedet for raa sekunder.
- Tastatur over bunn-sheets: KAV behavior "height" er doedt med
  edge-to-edge paa Android — byttet til "padding" i program-pickerne og
  legg-til-oevelse-modalen (faa soeketreff ble helt skjult).
- Komma og punktum fungerer begge som desimalskilletegn for vekt
  (parseDecimal-helper; norske tastaturer gir komma).
- Vektforslag baseres naa paa FOERSTE arbeidssett fra forrige oekt, ikke
  siste (siste er ofte et lettere back-off-sett) — og warmup-sett
  ekskluderes eksplisitt, som ogsaa fikser at oppvarmingsrampens sett
  kunne blitt «Sist»-forslaget.

Allerede levert i denne branchen (tester kjoerte eldre versjon):
kalender-exit-knapp, «Avslutt oekt» i bunnen + auto-prompt, «Fortsett
oekt»-kort paa hjem, splitting av program-supersett, chip-crop-fiksen.

+++ src/components/modals/PlateCalcModal.tsx
@@ -2,6 +2,7 @@
 import React, { useState, useEffect, useCallback } from "react";
 import { View, Text, Pressable, Modal, ScrollView, TextInput, Alert } from "react-native";
 import { useTheme } from "../../theme";
+import { parseDecimal } from "../../format";
 import { useI18n } from "../../i18n";
 import { useWeightUnit } from "../../units";
 import {
@@ -70,7 +71,7 @@ export default function PlateCalcModal({ visible, onClose, weightStr, exerciseId
 
   async function handleAddCustomBar() {
     const name = newBarName.trim();
-    const weightVal = parseFloat(newBarWeight);
+    const weightVal = parseDecimal(newBarWeight);
     if (!name) return;
     if (!Number.isFinite(weightVal) || weightVal < 0) return;
 
@@ -119,7 +120,7 @@ export default function PlateCalcModal({ visible, onClose, weightStr, exerciseId
   const selectedBar = allBars.find((b) => b.id === selectedBarId) ?? BUILT_IN_BARS[0];
   const barKg = selectedBar.kg;
 
-  const displayVal = parseFloat(weightStr);
+  const displayVal = parseDecimal(weightStr);
   const targetKg = Number.isFinite(displayVal) ? wu.toKg(displayVal) : 0;
   const result = calculatePlates(targetKg, barKg, gymPlates ?? undefined);
